chore: migrate to node:test
