+++ e2e/vite2/e2e-vite2.test.mjs
@@ -1,4 +1,4 @@
-import { runTests } from "../e2e";
+import { runTests } from "../e2e.mjs";
 
 runTests(2, async () => ({
   vite: await import("vite"),
