+++ e2e/vite3/e2e-vite3.test.mjs
@@ -1,6 +1,6 @@
-import { runTests } from "../e2e";
+import { runTests } from "../e2e.mjs";
 
-runTests(4, async () => ({
+runTests(3, async () => ({
   vite: await import("vite"),
   vitePluginLegacy: (await import("@vitejs/plugin-legacy")).default,
   vitePluginTopLevelAwait: (await import("vite-plugin-top-level-await")).default
