+++ e2e/vite4/e2e-vite4.test.mjs
@@ -1,8 +1,7 @@
-import { runTests } from "../e2e";
+import { runTests } from "../e2e.mjs";
 
-runTests(7, async () => ({
+runTests(4, async () => ({
   vite: await import("vite"),
-  // @ts-expect-error: @vitejs/plugin-legacy v7.0.0 doesn't have type export
   vitePluginLegacy: (await import("@vitejs/plugin-legacy")).default,
   vitePluginTopLevelAwait: (await import("vite-plugin-top-level-await")).default
 }));
