+++ e2e/vite7/e2e-vite7.test.mjs
@@ -1,6 +1,6 @@
-import { runTests } from "../e2e";
+import { runTests } from "../e2e.mjs";
 
-runTests(5, async () => ({
+runTests(7, async () => ({
   vite: await import("vite"),
   vitePluginLegacy: (await import("@vitejs/plugin-legacy")).default,
   vitePluginTopLevelAwait: (await import("vite-plugin-top-level-await")).default
