--- jest.config.js
@@ -1,20 +1,0 @@
-/** @type {import('ts-jest/dist/types').JestConfigWithTsJest} */
-module.exports = {
-  verbose: true,
-  transform: {
-    "^.+\\.ts?$": [
-      "ts-jest",
-      {
-        useESM: true,
-        tsconfig: "tsconfig.jest.json"
-      }
-    ]
-  },
-  extensionsToTreatAsEsm: [".ts"],
-  moduleNameMapper: {
-    "^(\\.{1,2}/.*)\\.js$": "$1"
-  },
-  testEnvironment: "node",
-  setupFilesAfterEnv: ["jest-extended/all"],
-  testPathIgnorePatterns: ["<rootDir>/dist/", "<rootDir>/node_modules/"]
-};
