+++ package.json
@@ -29,7 +29,7 @@
   ],
   "scripts": {
     "build": "tsc",
-    "test": "cross-env NODE_NO_WARNINGS=1 NODE_OPTIONS=--experimental-vm-modules jest --verbose --coverage --forceExit",
+    "test": "node src/wasm-parser.spec.mjs && node e2e/vite2/e2e-vite2.test.mjs && node e2e/vite3/e2e-vite3.test.mjs && node e2e/vite4/e2e-vite4.test.mjs && node e2e/vite5/e2e-vite5.test.mjs && node e2e/vite6/e2e-vite6.test.mjs && node e2e/vite7/e2e-vite7.test.mjs",
     "format": "prettier --write '**/*.{mjs,cjs,js,mts,cts,ts}'",
     "pretest": "yarn --cwd e2e && playwright install firefox && playwright install chromium"
   },
@@ -37,23 +37,18 @@
   "devDependencies": {
     "@babel/core": "^7.26.0",
     "@babel/preset-env": "^7.26.0",
-    "@jest/types": "^29.6.3",
     "@jsona/openapi": "^0.2.5",
     "@syntect/wasm": "^0.0.4",
     "@types/express": "^5.0.0",
-    "@types/jest": "^29.5.14",
     "@wasm-fmt/gofmt": "0.7.3",
-    "cross-env": "^7.0.3",
     "cz-conventional-changelog": "^3.3.0",
     "express": "^4.21.2",
-    "jest": "^29.7.0",
-    "jest-extended": "^4.0.2",
     "mime": "^4.0.6",
     "playwright": "1.23.3",
     "playwright-core": "1.23.3",
     "prettier": "^3.4.2",
     "terser": "^5.37.0",
-    "ts-jest": "^29.2.5",
+    "test": "^3.3.0",
     "typescript": "^5.7.2",
     "vite": "7",
     "wait-port": "^1.1.0"
