+++ test-helpers.mjs
@@ -1,0 +1,17 @@
+const nodeVersion = Number(process.versions.node.split(".")[0]);
+const mod = nodeVersion >= 16 ? await import("node:test") : (await import("test")).default;
+const { describe, it, before, after } = mod;
+
+if (typeof before !== "function") {
+  throw new Error("'before' should be a function");
+}
+if (typeof after !== "function") {
+  throw new Error("'after' should be a function");
+}
+if (typeof describe !== "function") {
+  throw new Error("'describe' should be a function");
+}
+if (typeof it !== "function") {
+  throw new Error("'it' should be a function");
+}
+export { describe, it, before, after };
