+++ tsconfig.json
@@ -27,6 +27,7 @@
   "exclude": [
     "node_modules",
     "dist",
-    "**/*.test.ts"
+    "**/*.spec.mjs",
+    "**/*.test.mjs"
   ]
 }
